# Improve caching for `theme.json` processing