--- a/src/wp-includes/class-wp-theme-json-resolver.php
+++ b/src/wp-includes/class-wp-theme-json-resolver.php
@@ -207,7 +207,9 @@ class WP_Theme_JSON_Resolver {
 			return true;
 		}
 
-		static::$blocks_cache[ $origin ] = $blocks;
+		foreach( $blocks as $block_name => $block_type ) {
+			static::$blocks_cache[ $origin ][ $block_name ] = true;
+		}
 
 		return false;
 	}