--- a/src/wp-includes/class-wp-theme-json-resolver.php
+++ b/src/wp-includes/class-wp-theme-json-resolver.php
@@ -193,9 +193,8 @@ class WP_Theme_JSON_Resolver {
 	 * @since 6.1.0
 	 *
 	 * @param string $origin Data source for which to cache the blocks.
-	                         Valid values are 'core', 'blocks', 'theme', and 'user'.
-	 *
-	 * @return boolean True on success, false otherwise.
+	 *                       Valid values are 'core', 'blocks', 'theme', and 'user'.
+	 * @return bool True on success, false otherwise.
 	 */
 	protected static function has_same_registered_blocks( $origin ) {
 		$registry = WP_Block_Type_Registry::get_instance();