@oandregal 2022-10-07
This is now in a good enough state we could consider merging, though there are some places where we can squeeze some more ms. I still need to fix tests as well.