@spacedmonkey 2022-10-10
@oandregal Can you please add some unit tests for this change. 101 lines change to a very important class, should include some unit tests.