@spacedmonkey 2022-10-10
<img width="593" alt="Screenshot 2022-10-10 at 11 18 50" src="https://user-images.githubusercontent.com/237508/194844896-8ef8ff56-fb9e-40c6-aa8b-8d5e0448baf7.png">

Still seeing lots of calls to remove keys in from schema.