@oandregal 2022-10-10
@ockham this is now ready. The numbers are back to 6.0 levels. I've updated the issue description with the performance results of calls to the main APIs people use to interact with `theme.json` data.

![image](https://user-images.githubusercontent.com/583546/194965304-ba466cca-39ec-470d-827d-1bcbc07b5501.png)

There are opportunities to squeeze a bit more ms, specially in translation (I tried, but tests failed for reasons I didn't understand and thought it was too risky for now, can prepare a follow-up PR if I find a way). More opportunities in reducing the number of calls to `WP_Theme_JSON_Resolver` methods, by inspecting higher-up the call stack (I didn't understand some of these places well enough to try). Speeding up the lower down (`WP_Theme_JSON` methods) parts seems a bit more complex, though we should try as well.