@oandregal 2022-10-10
> Still seeing lots of calls to remove keys in from schema.

There's value in reducing the number of times a function is called. This is the performance results for this function (I shared the xdebug profiles in the issue description for anyone to visualize with their tool of choice):

| Method                                                                        | WordPress 6.0     | WordPress trunk (rev 54449) | This PR          |
| --------------------------------------------------------------------- | ----------------------- | --------------------------- | ---------------- |
| `WP_Theme_JSON::remove_keys_not_in_schema` | 1 382 100ns / 249 calls = 5550ns per call  | 180 317 170ns / 33198 calls = 5431ns per call | 6 115 350ns / 1361 calls = 4493ns per call |

The times it takes the operation in the three setups I've tried does not indicate this is a bottleneck: it takes about the same amount of time per call. Though the number of times it's called indicates the issue is higher-up.

This is what I saw when zooming out a bit (call hierarchy for one path, via kcachegrind):

| WordPress 6.0     | WordPress trunk (rev 54449) | This PR          |
| ----------------------- | --------------------------- | ---------------- |
| ![Captura de ecrã de 2022-10-11 01-29-06](https://user-images.githubusercontent.com/583546/194967639-7f25efae-64cf-4a73-9b7a-9230d75bfa77.png)  | ![Captura de ecrã de 2022-10-11 01-28-38](https://user-images.githubusercontent.com/583546/194967644-c1b372c1-e31c-40fd-a9ef-35ce91c2e55c.png) | ![Captura de ecrã de 2022-10-11 01-26-32](https://user-images.githubusercontent.com/583546/194967652-1364b155-3250-421b-9c99-3747d5455888.png) |

My point is that `remove_keys_not_in_schema` is a leaf function that we can look at optimizing, though the nature of the issue was that we had to remove the cache we had, so it's only reasonable that some methods are called more times. The bigger gains in this case were (still are?) in bringing back that cache we had selectively.