@mukeshpanchal27 2022-10-11
@oandregal Please check https://github.com/WordPress/wordpress-develop/pull/3418#pullrequestreview-1136795378