@spacedmonkey 2022-10-11
Unit tests where would be useful.