@spacedmonkey 2022-10-11
To get a change this big merged in at the last minute, then unit tests are must. We do not want a regression at this late stage.