@spacedmonkey 2022-10-11
There are not unit tests for the following methods. 

- get_core_data
- has_same_registered_blocks
- get_theme_data
- get_user_data

I don't think we have enough coverage to merge this change. Tests to prove that methods above do as expected. Check cleaning caches was as expected and code is not being run again after cache working as expected. 

There are basically no changes to WordPress that should not come along with a unit test. Something this important NEEDS them before the mege.