@oandregal 2022-10-11
I'm still investigating this for further improvement and have created https://github.com/WordPress/wordpress-develop/pull/3441 as a follow-up.